+++ src/shared/constants/sites.js
@@ -8,10 +8,10 @@
 
 /**
  * Configuration for supported coding sites
- * Currently supports: LeetCode
+ * Currently supports: LeetCode, TakeUForward, GeeksForGeeks
  * Architecture allows easy addition of new sites in the future
  */
-const LEETCODE_CONFIG = {
+const MONACO_CONFIG = {
   editor: 'monaco',
   selectors: ['.monaco-editor'],
   waitStrategy: 'mutation-observer',
@@ -19,17 +19,40 @@ const LEETCODE_CONFIG = {
     namespace: 'monaco',
     getEditors: () => window.monaco?.editor?.getEditors?.() || [],
     keyCode: {
-      Tab: 3  // monaco.KeyCode.Tab
+      Tab: 2 // monaco.KeyCode.Tab
+    }
+  }
+};
+const LEETCODE_CONFIG = {
+  ...MONACO_CONFIG,
+};
+
+const TAKEUFORWARD_CONFIG = {
+  ...MONACO_CONFIG,
+};
+
+const ACE_CONFIG = {
+  editor: 'ace',
+  selectors: ['.ace_editor'],
+  waitStrategy: 'mutation-observer',
+  editorApi: {
+    namespace: 'ace',
+    getEditor: () => {
+      const el = document.querySelector('.ace_editor');
+      return el && window.ace?.edit?.(el);
     }
   }
 };
 
+const GFG_CONFIG = {
+  ...ACE_CONFIG,
+};
+
 export const SITE_CONFIGS = {
   'leetcode.com': LEETCODE_CONFIG,
-  'leetcode.cn': LEETCODE_CONFIG
-  // Future sites can be added here:
-  // 'geeksforgeeks.org': { ... },
-  // 'takeuforward.org': { ... }
+  'leetcode.cn': LEETCODE_CONFIG,
+  'takeuforward.org': TAKEUFORWARD_CONFIG,
+  'geeksforgeeks.org': GFG_CONFIG,
 };
 
 /**